+++ src/kschltz/agent/stream/wiring.clj
@@ -3,11 +3,14 @@
   (:require [integrant.core :as ig]
             [kschltz.agent.stream.bus :as bus]
             [kschltz.agent.stream.plugin :as stream.plugin]
+            [kschltz.agent.stream.store-bus :as store-bus]
             [malli.core :as m]))
 
 (def StreamBusConfig
   [:map
-   [:enabled? {:optional true} :boolean]])
+   [:enabled? {:optional true} :boolean]
+   [:impl {:optional true} [:enum :memory :store]]
+   [:store {:optional true} :any]])
 
 (defn- assert-malli!
   [key schema value]
@@ -22,7 +25,14 @@
 
 (defmethod ig/init-key :lateralus/stream-bus [_ opts]
   (when-not (false? (:enabled? opts))
-    (bus/create-bus)))
+    (case (keyword (or (:impl opts) :memory))
+      :memory (bus/create-bus)
+      :store (if-let [engine (:store opts)]
+               (store-bus/store-bus engine)
+               (throw (ex-info "stream-bus :impl :store requires :store"
+                               {:opts opts})))
+      (throw (ex-info "Unknown :lateralus/stream-bus :impl"
+                      {:impl (:impl opts)})))))
 
 (defmethod ig/init-key :lateralus/stream-plugin [_ {:keys [bus]}]
   (stream.plugin/stream-plugin bus))